+++ src/core/db/model.py
@@ -4,8 +4,9 @@ from geoalchemy2.types import Geography
 from sqlalchemy import BigInteger, Boolean, Column, Date, DateTime, Float, ForeignKey, Integer, String, Table, Text
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import backref, relationship
+from werkzeug.security import check_password_hash, generate_password_hash
 
-from core.mixins import ExtraUserMixin
+
 from src.core.db.db import Base
 
 
@@ -81,7 +82,7 @@ class Role(Base, RoleMixin):
         return self.name
 
 
-class Staff(Base, UserMixin, ExtraUserMixin):
+class Staff(Base, UserMixin):
     """Модель персонала"""
 
     first_name = Column(String(255))
@@ -91,3 +92,18 @@ class Staff(Base, UserMixin, ExtraUserMixin):
     password = Column(String(255))
     active = Column(Boolean())
     roles = relationship("Role", secondary=roles_users, backref=backref("users", lazy="dynamic"))
+
+    def has_role(self, *args):
+        return set(args).issubset({role.name for role in self.roles})
+
+    def get_id(self):
+        return self.id
+
+    def __unicode__(self):
+        return self.login
+
+    def set_password(self, password):
+        self.password = generate_password_hash(password)
+
+    def check_password(self, password):
+        return check_password_hash(self.password, password)
