--- src/core/mixins.py
@@ -1,18 +1,0 @@
-from werkzeug.security import check_password_hash, generate_password_hash
-
-
-class ExtraUserMixin:
-    def has_role(self, *args):
-        return set(args).issubset({role.name for role in self.roles})
-
-    def get_id(self):
-        return self.id
-
-    def __unicode__(self):
-        return self.login
-
-    def set_password(self, password):
-        self.password = generate_password_hash(password)
-
-    def check_password(self, password):
-        return check_password_hash(self.password, password)
